requirements.txt
sounddevice
scipy
assemblyai
google-generativeai
